chore (refactor): move trekko_test_utils.dart class to utils
