+++ test/profile_test.dart
@@ -6,7 +6,7 @@ import 'package:trekko_backend/model/profile/profile.dart';
 import 'package:trekko_backend/model/profile/question_type.dart';
 import 'package:test/test.dart';
 
-import 'trekko_test_utils.dart';
+import 'utils/trekko_test_utils.dart';
 
 void main() {
   late Trekko trekko;
