+++ test/tracking/cache_tracking_test.dart
@@ -8,7 +8,7 @@ import 'package:fling_units/fling_units.dart';
 import 'package:isar/isar.dart';
 import 'package:test/test.dart';
 
-import '../trekko_test_utils.dart';
+import '../utils/trekko_test_utils.dart';
 import '../utils/tracking_test_util.dart';
 
 void main() {
